+++ src/FormulationCompact.h
@@ -1,7 +1,6 @@
 /** 
  * @class FormulationCompact
  ** An object from this class has the variable and constraints defining a UCP problem.
- *  todo : improve the printing to get it with cout ?
 */
 
 #ifndef FormulationCompact_H
@@ -28,34 +27,42 @@ class FormulationCompact
 
     public : 
 
-    FormulationCompact(InstanceUCP *instance, SCIP *scip);
-    SCIP_RETCODE create_variables();
-    SCIP_RETCODE create_constraints();
+        /* constructor */
+        FormulationCompact(InstanceUCP *instance, SCIP *scip);
+
+        /* create the variables, add them to scip and store them */
+        SCIP_RETCODE create_variables();
+
+        /* create the constraints, add them to scip problem and store them */
+        SCIP_RETCODE create_constraints();
+
+        
+        /* gets */
+
+        SCIP* get_scip_pointer();
+
+        std::vector< std::vector< SCIP_VAR* >> get_variable_u();
+        std::vector< std::vector< SCIP_VAR* >> get_variable_x();
+        std::vector< std::vector< SCIP_VAR* >> get_variable_p();
+
+        std::vector< std::vector< SCIP_CONS* >> m_constraint_min_up_time;
 
 
-    SCIP* get_scip_pointer();
-    
-
-    std::vector< std::vector< SCIP_VAR* >> get_variable_u();
-    std::vector< std::vector< SCIP_VAR* >> get_variable_x();
-    std::vector< std::vector< SCIP_VAR* >> get_variable_p();
-
-    std::vector< std::vector< SCIP_CONS* >> m_constraint_min_up_time;
 
     private:
 
-    SCIP *p_scip;
-    InstanceUCP *p_ucp_instance;
+        SCIP *p_scip;
+        InstanceUCP *p_ucp_instance;
 
-    std::vector< std::vector< SCIP_VAR* >> m_variable_u;
-    std::vector< std::vector< SCIP_VAR* >> m_variable_x;
-    std::vector< std::vector< SCIP_VAR* >> m_variable_p;
+        std::vector< std::vector< SCIP_VAR* >> m_variable_u;
+        std::vector< std::vector< SCIP_VAR* >> m_variable_x;
+        std::vector< std::vector< SCIP_VAR* >> m_variable_p;
 
-    std::vector< SCIP_CONS* > m_constraint_demand;
-    std::vector< std::vector< SCIP_CONS* >> m_constraint_production;
-    // std::vector< std::vector< SCIP_CONS* >> m_constraint_min_up_time;
-    std::vector< std::vector< SCIP_CONS* >> m_constraint_min_down_time;
-    std::vector< std::vector< SCIP_CONS* >> m_constraint_startup;
+        std::vector< SCIP_CONS* > m_constraint_demand;
+        std::vector< std::vector< SCIP_CONS* >> m_constraint_production;
+        // std::vector< std::vector< SCIP_CONS* >> m_constraint_min_up_time;
+        std::vector< std::vector< SCIP_CONS* >> m_constraint_min_down_time;
+        std::vector< std::vector< SCIP_CONS* >> m_constraint_startup;
      
 
 };
