--- src/FormulationMaster.cpp
@@ -1,170 +1,0 @@
-/* Implements the class FormulationMaster */
-
-/* standart includes */
-#include <vector>
-#include <iostream>
-#include <string>
-#include <fstream>
-#include <sstream>
-
-/* SCIP includes */
-#include <scip/scipdefplugins.h>
-#include <scip/scip.h>
-
-/* user includes */
-#include "InstanceUCP.h"
-#include "FormulationMaster.h"
-#include "VariableMaster.h"
-#include "ProductionPlan.h"
-
-/* namespace */
-using namespace std;
-
-/** constructor */
-FormulationMaster::FormulationMaster( InstanceUCP* instance, SCIP* scip_master)
-{
-    p_instance = instance;
-    p_scip_master = scip_master;
-
-
-    // creating the constraints
-    ostringstream current_cons_name;
-
-    // convexity constraint
-    m_convexity_constraint = new SCIP_CONS;
-
-    current_cons_name.str("");
-    current_cons_name << "convexity_constraint";
-    SCIPcreateConsLinear( p_scip_master, &m_convexity_constraint, current_cons_name.str().c_str(), 0, NULL, NULL,
-			    1.0,   // lhs 
-			    1.0,   // rhs 
-			    true,  /* initial */
-			    false, /* separate */
-			    true,  /* enforce */
-			    true,  /* check */
-			    true,  /* propagate */
-			    false, /* local */
-			    true,  /* modifiable */
-			    false, /* dynamic */
-			    false, /* removable */
-			    false  /* stickingatnode */ );
-    SCIPaddCons( p_scip_master, m_convexity_constraint );
-
-
-
-    // Complicating constraints
-    int number_time_step( p_instance->get_time_steps_number() );
-    m_complicating_constraints.resize( number_time_step );
-    vector<int> demand( p_instance->get_demand() );
-    for( int i_time_step = 0; i_time_step < number_time_step; i_time_step ++)
-    {
-        SCIP_CONS* demand_constraint_t;
-
-        current_cons_name.str("");
-        current_cons_name << "demand_constraint_" << i_time_step;
-        SCIPcreateConsLinear( p_scip_master, &demand_constraint_t, current_cons_name.str().c_str(), 0, NULL, NULL,
-			    demand[i_time_step],   // lhs 
-			    + SCIPinfinity( p_scip_master ),   // rhs 
-			    true,  /* initial */
-			    false, /* separate */
-			    true,  /* enforce */
-			    true,  /* check */
-			    true,  /* propagate */
-			    false, /* local */
-			    true,  /* modifiable */
-			    false, /* dynamic */
-			    false, /* removable */
-			    false  /* stickingatnode */ );
-        SCIPaddCons( p_scip_master, demand_constraint_t );
-        m_complicating_constraints[i_time_step] = demand_constraint_t;
-    }
-
-}
-
-
-
-
-FormulationMaster::~FormulationMaster()
-{}
-
-
-void FormulationMaster::addColumn( VariableMaster* variable_to_add )
-{
-    
-    SCIP_VAR* p_variable = variable_to_add->get_variable_pointer();
-
-    // add column to convexity constraint
-    SCIPaddCoefLinear(p_scip_master,
-                m_convexity_constraint,
-                p_variable,  /* variable to add */
-                1.);         /* coefficient */        
-    
-    // add column to demand constraints
-    int number_time_step( p_instance->get_time_steps_number() );
-    vector<vector<double>> quantity_plan = variable_to_add->get_production_plan()->get_quantity_plan();
-    for( int i_time_step = 0; i_time_step < number_time_step; i_time_step++ )
-    {
-        double quantity_plan_t(0);
-        for( int i_unit = 0; i_unit < p_instance->get_units_number(); i_unit++ )
-        {
-            quantity_plan_t += quantity_plan[i_unit][i_time_step];
-        }
-        SCIPaddCoefLinear(p_scip_master,
-            m_complicating_constraints[i_time_step],
-            p_variable,  /* variable to add */
-            quantity_plan_t);                                    /* coefficient */
-    }
-
-    m_vector_columns.push_back( variable_to_add );
-
-
-    return;
-}
-
-
-vector<int> FormulationMaster::get_dual_sol()
-{
-    vector<int> dual_sol;
-    int number_time_steps = p_instance->get_time_steps_number();
-    dual_sol.resize(number_time_steps);
-    for(int i_time_step = 0; i_time_step < number_time_steps; i_time_step ++)
-    {
-        int dual_sol_t( SCIPgetDualsolLinear(p_scip_master, m_complicating_constraints[i_time_step] )); 
-        cout << "Dual cost for the demand constraint " << i_time_step << " : " << dual_sol_t << "\n";
-        dual_sol[i_time_step] = dual_sol_t;
-    }
-}
-
-
-ProductionPlan* FormulationMaster::get_production_plan_from_solution()
-{
-
-    int number_columns( m_vector_columns.size());
-    
-    // create empty plan
-    ProductionPlan* production_plan = new ProductionPlan( p_instance );
-    production_plan->empty_plan();
-
-    // now add the columns values to the plan
-    SCIP_SOL *solution = SCIPgetBestSol( p_scip_master );
-
-    for(int i_column = 0; i_column < number_columns; i_column ++)
-    {
-        // get the plan
-        VariableMaster* current_variable = m_vector_columns[i_column];
-        ProductionPlan* current_plan = current_variable->get_production_plan();
-        
-        // add the value to the new plan
-        SCIP_Real coefficient_column( SCIPgetSolVal( p_scip_master, solution, current_variable->get_variable_pointer()));
-        production_plan->add_column_values( current_plan, coefficient_column);
-        production_plan->computeCost();
-    }
-
-    return( production_plan);
-}
-
-
-
-
-
-
